move listPlugins main code
