+++ packages/core/src/utils/__tests__/load-plugin-windows.test.ts
@@ -1,4 +1,4 @@
-import loadPlugin from "../load-plugins";
+import { loadPlugin } from "../load-plugins";
 import { dummyLog } from "../logger";
 
 const logger = dummyLog();
